+++ StateMachine/GraphableStateMachineSchema.swift
@@ -3,26 +3,32 @@ import Foundation
 
 /// A type, preferably an `enum`, representing states or events in
 /// a State Machine.  Used by `GraphableStateMachineSchema` to create
-/// state machine graphs in the DOT graph description language [1].
+/// state machine graphs in the [DOT graph description language][DOT].
 ///
 /// Provides an array of items (states or events) used in a graph, and
 /// a textual representation of each item, used when assigning labels
 /// to graph elements.
 ///
-///   [1]: http://en.wikipedia.org/wiki/DOT_%28graph_description_language%29
+///   [DOT]: http://en.wikipedia.org/wiki/DOT_%28graph_description_language%29
 public protocol DOTLabelable {
 
     /// A textual representation of `self`, suitable for creating labels
-    /// in a graph.
+    /// in a graph.  If not provided, default string representation is used.
     var DOTLabel: String { get }
 
     /// An array of items of a given type (states or events) used in a graph.
     static var DOTLabelableItems: [Self] { get }
 }
 
+public extension DOTLabelable {
+    var DOTLabel: String {
+        return "\(self)"
+    }
+}
 
-/// A state machine schema with a graph of the state machine in the DOT graph
-/// description language [1].
+
+/// A state machine schema with a graph of the state machine in the [DOT graph
+/// description language][DOT].
 ///
 /// Requires `State` and `Event` types to conform to the `DOTLabelable`
 /// protocol.
@@ -34,11 +40,11 @@ public protocol DOTLabelable {
 /// For more information about state machine schemas, see
 /// `StateMachineSchemaType` documentation.
 ///
-///   [1]: http://en.wikipedia.org/wiki/DOT_%28graph_description_language%29
+///   [DOT]: http://en.wikipedia.org/wiki/DOT_%28graph_description_language%29
 public struct GraphableStateMachineSchema<A: DOTLabelable, B: DOTLabelable, C>: StateMachineSchemaType {
-    typealias State = A
-    typealias Event = B
-    typealias Subject = C
+    public typealias State = A
+    public typealias Event = B
+    public typealias Subject = C
 
     public let initialState: State
     public let transitionLogic: (State, Event) -> (State, (Subject -> ())?)?
@@ -73,15 +79,15 @@ public struct GraphableStateMachineSchema<A: DOTLabelable, B: DOTLabelable, C>: 
     /// `MyProject/InboxViewController.swift` and pass `"Inbox.dot"` as
     /// a filepath, the diagram will be saved as `MyProject/Inbox.dot`.
     ///
-    /// Files in the DOT format [1] can be viewed in different applications,
-    /// e.g. Graphviz [2].
+    /// Files in the [DOT format][DOT] can be viewed in different applications,
+    /// e.g. [Graphviz][].
     ///
-    ///   [1]: https://developer.apple.com/library/mac/qa/qa1361/_index.html
-    ///   [2]: http://www.graphviz.org/
-    public func saveDOTDigraphIfRunningInSimulator(#filepathRelativeToCurrentFile: String, file: String = __FILE__) {
+    ///   [DOT]: http://en.wikipedia.org/wiki/DOT_%28graph_description_language%29
+    ///   [Graphviz]: http://www.graphviz.org/
+    public func saveDOTDigraphIfRunningInSimulator(filepathRelativeToCurrentFile filepathRelativeToCurrentFile: String, file: String = __FILE__) throws {
         if TARGET_IPHONE_SIMULATOR == 1 {
-            let filepath = file.stringByDeletingLastPathComponent.stringByAppendingPathComponent(filepathRelativeToCurrentFile)
-            DOTDigraph.writeToFile(filepath, atomically: true, encoding: NSUTF8StringEncoding, error: nil)
+            let filepath = ((file as NSString).stringByDeletingLastPathComponent as NSString).stringByAppendingPathComponent(filepathRelativeToCurrentFile)
+            try DOTDigraph.writeToFile(filepath, atomically: true, encoding: NSUTF8StringEncoding)
         }
     }
     #endif
@@ -91,7 +97,7 @@ public struct GraphableStateMachineSchema<A: DOTLabelable, B: DOTLabelable, C>: 
         let events = Event.DOTLabelableItems
 
         var stateIndexesByLabel: [String: Int] = [:]
-        for (i, state) in enumerate(states) {
+        for (i, state) in states.enumerate() {
             stateIndexesByLabel[label(state)] = i + 1
         }